refactor visitor and linter; partially resolve review
