+++ packages/core/src/graphql/detect-graphql.ts
